fix: 둘러보기 컴포넌트 분리
